Extract remote client health state
