+++ packages/client/src/remote-client.ts
@@ -9,12 +9,10 @@ import type {
   PickRawFields,
   RawQuery,
   StatusEvent,
-  TopicRuntimeHealth,
   TopicDefinitions,
   TopicRow,
   ValidateLiveQuery,
   ViewServerConfig,
-  ViewServerHealth,
   ViewServerHealthSummaryRow,
   ViewServerHealthTopicRow,
   ViewServerRuntimeError,
@@ -35,7 +33,6 @@ import {
   type ViewServerWireLiveQuery,
 } from "@view-server/protocol";
 import { Context, Effect, Exit, Layer, ManagedRuntime, Scope, Stream } from "effect";
-import { AtomRef } from "effect/unstable/reactivity";
 import { RpcClient, RpcSerialization } from "effect/unstable/rpc";
 import type { RpcClientError } from "effect/unstable/rpc/RpcClientError";
 import type {
@@ -43,6 +40,7 @@ import type {
   ViewServerLiveEvent,
   ViewServerLiveSubscription,
 } from "./live-client";
+import { makeRemoteHealthState } from "./remote-health";
 import { makeRemoteSubscription } from "./remote-subscription";
 
 export type ViewServerRemoteClientError = ViewServerRuntimeError | ViewServerTransportError;
@@ -69,38 +67,6 @@ const transportError = (error: Error): ViewServerTransportError => ({
   _tag: "ViewServerTransportError",
   code: "TransportError",
   message: error.message,
-});
-
-function typedHealthTopics<Topics extends TopicDefinitions>(
-  topics: Record<string, TopicRuntimeHealth>,
-): ViewServerHealth<Topics>["engine"]["topics"];
-function typedHealthTopics(
-  topics: Record<string, TopicRuntimeHealth>,
-): Record<string, TopicRuntimeHealth> {
-  return topics;
-}
-
-const topicHealthFromRow = (
-  existing: TopicRuntimeHealth,
-  row: ViewServerHealthTopicRow,
-): TopicRuntimeHealth => ({
-  status: row.status === "stopping" ? existing.status : row.status,
-  rowCount: row.rowCount,
-  liveRowCount: row.liveRowCount,
-  deletedRowCount: row.deletedRowCount,
-  version: row.version,
-  lastMutationAt: row.lastMutationAt,
-  mutationsPerSecond: row.mutationsPerSecond,
-  rowsPerSecond: row.rowsPerSecond,
-  pendingMutationBatches: row.pendingMutationBatches,
-  activeViews: row.activeViews,
-  activeSubscriptions: row.activeSubscriptions,
-  queuedEvents: row.queuedEvents,
-  maxQueueDepth: row.maxQueueDepth,
-  backpressureEvents: row.backpressureEvents,
-  memoryBytes: row.memoryBytes,
-  tombstoneCount: row.tombstoneCount,
-  compactionPending: row.compactionPending,
 });
 
 export const mapViewServerRemoteError = (
@@ -187,124 +153,13 @@ export const makeViewServerClient: <const Topics extends TopicDefinitions>(
   const initialHealth = yield* cleanupOnConstructionFailure(
     healthRpc().pipe(Effect.flatMap((next) => viewServerDecodeHealth(config, next))),
   );
-  const health = AtomRef.make(initialHealth);
+  const remoteHealth = makeRemoteHealthState(initialHealth);
   const subscriptionBufferSize = options.subscriptionBufferSize ?? 1_024;
   const clientScope = yield* Scope.make("parallel");
 
-  const updateHealthSummaryRef = (event: ViewServerLiveEvent<ViewServerHealthSummaryRow<Topics>>) =>
-    Effect.sync(() => {
-      const applySummaryRow = (row: ViewServerHealthSummaryRow<Topics>) => {
-        health.update((current) => ({
-          ...current,
-          status: row.runtimeStatus,
-        }));
-      };
-      if (event.type === "snapshot") {
-        for (const row of event.rows) {
-          applySummaryRow(row);
-        }
-      }
-      if (event.type === "delta") {
-        for (const operation of event.operations) {
-          if (operation.type === "insert" || operation.type === "update") {
-            applySummaryRow(operation.row);
-          }
-        }
-      }
-    });
-
-  const updateHealthTopicRef = (
-    event: ViewServerLiveEvent<ViewServerHealthTopicRow<Extract<keyof Topics, string>>>,
-  ) =>
-    Effect.sync(() => {
-      if (event.type === "snapshot") {
-        health.update((current) => {
-          const topics: Record<string, TopicRuntimeHealth> = { ...current.engine.topics };
-          for (const row of event.rows) {
-            topics[row.id] = topicHealthFromRow(current.engine.topics[row.id], row);
-          }
-          return {
-            ...current,
-            engine: {
-              topics: typedHealthTopics<Topics>(topics),
-            },
-          };
-        });
-      }
-      if (event.type === "delta") {
-        health.update((current) => {
-          const topics: Record<string, TopicRuntimeHealth> = { ...current.engine.topics };
-          for (const operation of event.operations) {
-            if (operation.type === "insert" || operation.type === "update") {
-              topics[operation.key] = topicHealthFromRow(
-                current.engine.topics[operation.row.id],
-                operation.row,
-              );
-            }
-          }
-          return {
-            ...current,
-            engine: {
-              topics: typedHealthTopics<Topics>(topics),
-            },
-          };
-        });
-      }
-    });
-
-  const updateLiveTopicHealth = <Topic extends Extract<keyof Topics, string>>(
-    topic: Topic,
-    update: (current: TopicRuntimeHealth) => TopicRuntimeHealth,
-  ) =>
-    Effect.sync(() => {
-      health.update((current) => {
-        const topics: Record<string, TopicRuntimeHealth> = {
-          ...current.engine.topics,
-          [topic]: update(current.engine.topics[topic]),
-        };
-        return {
-          ...current,
-          engine: {
-            topics: typedHealthTopics<Topics>(topics),
-          },
-        };
-      });
-    });
-
-  const updateSubscriptionCount = <Topic extends Extract<keyof Topics, string>>(
-    topic: Topic,
-    delta: 1 | -1,
-  ) =>
-    Effect.gen(function* () {
-      yield* updateLiveTopicHealth(topic, (current) => {
-        const activeSubscriptions = Math.max(0, current.activeSubscriptions + delta);
-        return {
-          ...current,
-          activeSubscriptions,
-        };
-      });
-      yield* Effect.sync(() => {
-        health.update((current) => ({
-          ...current,
-          transport: {
-            ...current.transport,
-            activeStreams: Math.max(0, current.transport.activeStreams + delta),
-            activeSubscriptions: Math.max(0, current.transport.activeSubscriptions + delta),
-          },
-        }));
-      });
-    });
-
   const close = Scope.close(clientScope, Exit.void).pipe(
     Effect.andThen(managedRuntime.disposeEffect),
-    Effect.andThen(
-      Effect.sync(() => {
-        health.update((current) => ({
-          ...current,
-          status: "stopping",
-        }));
-      }),
-    ),
+    Effect.andThen(remoteHealth.markStopping),
   );
 
   const streamToSubscription = <Row>(
@@ -337,8 +192,8 @@ export const makeViewServerClient: <const Topics extends TopicDefinitions>(
       viewServerDecodeLiveEvent<Topics, Topic, Row>(config, topic, wireQuery, event),
     );
     return yield* streamToSubscription(topic, stream, {
-      onOpen: updateSubscriptionCount(topic, 1),
-      onClose: updateSubscriptionCount(topic, -1),
+      onOpen: remoteHealth.updateSubscriptionCount(topic, 1),
+      onClose: remoteHealth.updateSubscriptionCount(topic, -1),
     });
   });
 
@@ -391,7 +246,7 @@ export const makeViewServerClient: <const Topics extends TopicDefinitions>(
         (event) => viewServerDecodeHealthSummaryEvent(config, event),
       );
       const subscription = yield* streamToSubscription(VIEW_SERVER_HEALTH_SUMMARY_TOPIC, stream);
-      const events = subscription.events.pipe(Stream.tap(updateHealthSummaryRef));
+      const events = subscription.events.pipe(Stream.tap(remoteHealth.updateHealthSummaryRef));
       return {
         events,
         close: subscription.close,
@@ -406,7 +261,7 @@ export const makeViewServerClient: <const Topics extends TopicDefinitions>(
       viewServerDecodeHealthTopicEvent(config, event),
     );
     const subscription = yield* streamToSubscription(VIEW_SERVER_HEALTH_TOPIC, stream);
-    const events = subscription.events.pipe(Stream.tap(updateHealthTopicRef));
+    const events = subscription.events.pipe(Stream.tap(remoteHealth.updateHealthTopicRef));
     return {
       events,
       close: subscription.close,
@@ -417,7 +272,7 @@ export const makeViewServerClient: <const Topics extends TopicDefinitions>(
     subscribe,
     subscribeHealthSummary,
     subscribeHealth,
-    health: health.map((value) => value),
+    health: remoteHealth.readonlyHealth,
     close,
   };
 });
